Fix remaining MSVC /W4 warnings: shadowing and narrowing
The Windows CI build revealed /W4 warnings the Clang -Wall -Wextra sweep
cannot see (C4456/C4458 member and local shadowing, C4005 macro
redefinition, C4244 float-to-int and int-to-Uint8 narrowing). Fix them
so /WX stays green:

- AudioFile.h: guard NOMINMAX against the CMake compile definition
- Sprite.h/.cpp, TileMap.h, Game2D.cpp: rename params and locals that
  shadow Object members (renderer, id) and SDLApp assetManager
- SkeletalAnim: rename the backward-pass loop index that shadows the
  forward-pass counter
- CinderfallState: rename the load() local that shadows State::maxFloor
- GameJuice/Breakout/BrickBreakerPlus: cast color components to Uint8
- Frogger: make the row-index constants int, not float

Verified locally with -Wshadow -Wconversion -Wshorten-64-to-32 and a
-Werror native build of the affected targets.

+++ Engine/Core/GameJuice.h
@@ -314,7 +314,7 @@ public:
 
     void render(SDL_Renderer* sdl, int offsetX = 0, int offsetY = 0) const {
         for (const Particle& p : ps_) {
-            const int a = (int)(255.0f * (p.life / p.maxLife));
+            const Uint8 a = (Uint8)(255.0f * (p.life / p.maxLife));
             SDL_SetRenderDrawColor(sdl, p.color.r, p.color.g, p.color.b, a);
             const int s = (int)p.size;
             SDL_Rect r = {(int)p.x + offsetX - s / 2,
